# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Visitas",
        "description": "API para manejar las visitas almacenadas en la base de datos.",
        "contact": {
            "email": "contacto@tusitio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/getVisit": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Obtener todas las visitas",
                "description": "Recupera todas las visitas almacenadas en la base de datos.",
                "operationId": "ef9eabf85d2b23ab9d6e5010f46daee7",
                "responses": {
                    "200": {
                        "description": "Lista de visitas obtenida con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Viviana Herrera"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "vivianaherrerahe@gmail.com"
                                            },
                                            "latitude": {
                                                "type": "string",
                                                "example": "4.7110"
                                            },
                                            "longitude": {
                                                "type": "string",
                                                "example": "-74.0721"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-11T00:54:16.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron registros.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron registros."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al obtener los datos."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/setVisits": {
            "post": {
                "tags": [
                    "Visits"
                ],
                "summary": "Crear una nueva visita",
                "description": "Este endpoint permite crear una nueva visita con los datos proporcionados.",
                "operationId": "db42fdf17aa2a04415c7a0c836d36fa5",
                "requestBody": {
                    "description": "Datos de la nueva visita.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juanperez@correo.com"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "4.7110"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "-74.0721"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Visita creada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Visita creada exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        },
                                        "message": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo name es obligatorio."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo email es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno al crear la visita.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al crear la visita"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/visitId/{id}": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Obtener una visita por ID",
                "description": "Este endpoint permite obtener los detalles de una visita existente por su ID.",
                "operationId": "c9a7db9a1c9d67f3c8a2a95d95a605e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la visita que se desea obtener.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visita encontrada con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Juan Pérez"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "juanperez@correo.com"
                                        },
                                        "latitude": {
                                            "type": "string",
                                            "example": "4.7110"
                                        },
                                        "longitude": {
                                            "type": "string",
                                            "example": "-74.0721"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-11T00:54:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-11T00:54:16.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visita no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Visita no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno al obtener la visita.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al obtener la visita"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/delete/{id}": {
            "delete": {
                "tags": [
                    "Visits"
                ],
                "summary": "Eliminar una visita",
                "description": "Este endpoint permite eliminar una visita existente utilizando su ID.",
                "operationId": "5b5ddfedca380a4536a1a1102ad8550b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la visita a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visita eliminada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Visita eliminada exitosamente!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visita no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Visita no encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno al eliminar la visita.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocurrió un error al eliminar la visita."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}